requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # For YOLOv8
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
albumentations>=1.3.0
tensorboard>=2.12.0 
transformers  # for TrOCR
pytesseract   # for baseline comparison